Add DependsOn removal and Intrinsic function transformation
Remove the DependsOn statements in nested stacks so that they have no dependency on a non-reachable resource in a parent stack. Furthermore the Fn::GetAtt usages which try to access external dependencies are updated to Refs. Those intrinsic function calls are moved from the nested stack to the parent stack and passed down as stack parameters.